[#9] .Refactor moved the quitChannel from websocket to Client
This avoids data flow from websocket to Client. 
Furthermore there is no need that the websocket has to notify the Client about a closed connection. The Client is anyway the only one closing the websocket connection hence he knows when the connection is closed and when the Clients worker threads can be stopped.