.Fix reworked approach to check health state
Using the information if there was a successful connection is not a good idea since one needs to create at least one connection beforehand. It is the better approach to find out actively if a connection to the peer is possible or not. 
Hence we now create a websocket and then send a ping control message over this connection to the peer.